+++ src/main/java/mops/authentication/account/keycloak/KeycloakConfig.java
@@ -1,4 +1,4 @@
-package mops.authentication.keycloak;
+package mops.authentication.account.keycloak;
 
 import org.keycloak.adapters.springboot.KeycloakSpringBootConfigResolver;
 import org.springframework.context.annotation.Bean;
